requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
bitarray==2.9.2
boto3==1.34.149
botocore==1.34.149
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
coverage==7.6.1
cytoolz==0.12.3
Django==5.0.5
django-cors-headers==4.3.1
django-webpack-loader==3.1.0
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
drf-spectacular==0.27.2
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
packaging==24.0
panzi-json-logic==1.0.1
parsimonious==0.10.0
protobuf==5.26.1
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
toolz==0.12.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
web3==6.18.0
websockets==12.0
yarl==1.9.4
